# feat: Phase 2 arena-based cache optimization foundation